+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Remove `cbor-smol` re-export as `serde`.
 - Make `ctap2::CtapMappingError` private.
 - Set `#[repr(u8)]` for `ctap2::Error` and implement `From<ctap2::Error>` for `u8`.
+- `ctap2::get_info`: Use `ByteArray<16>` instead of `Bytes<16>` for AAGUID.
 
 ## [0.6.0-rc.4] 2026-06-01